db/scripts/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
requests==2.31.0
influxdb==5.3.1
prometheus-client==0.19.0